backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-unfold==0.89.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
qrcode==8.2
reportlab==4.4.10
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
